Added GetErrorResponse so in the controller it calls this function instead of handling the logic if it is an error in the getresponse. Also moved the structs into utils/response.go
