+++ backend/controllers/users.go
@@ -17,13 +17,13 @@ func GetUsers(c *gin.Context) {
 
 	page, err := strconv.Atoi(pageStr)
 	if err != nil {
-		response := utils.GetResponse(nil, nil, http.StatusBadRequest, "Page parameter is not a number")
+		response := utils.GetErrorResponse(http.StatusBadRequest, "Page parameter is not a number")
 		c.JSON(http.StatusBadRequest, response)
 		return
 	}
 	limit, err := strconv.Atoi(limitStr)
 	if err != nil {
-		response := utils.GetResponse(nil, nil, http.StatusBadRequest, "Limit parameter is not a number")
+		response := utils.GetErrorResponse(http.StatusBadRequest, "Limit parameter is not a number")
 		c.JSON(http.StatusBadRequest, response)
 		return
 	}
@@ -32,19 +32,17 @@ func GetUsers(c *gin.Context) {
 	paginatedUsers, err := repositories.GetPaginatedUsers(page, limit)
 	if err != nil {
 		if errors.As(err, &paginationErr) {
-			response := utils.GetResponse(nil, nil, http.StatusBadRequest, paginationErr.Error())
+			response := utils.GetErrorResponse(http.StatusBadRequest, paginationErr.Error())
 			c.JSON(http.StatusBadRequest, response)
 			return
 		} else {
-			response := utils.GetResponse(nil, nil, http.StatusInternalServerError, "")
+			response := utils.GetErrorResponse(http.StatusInternalServerError, "")
 			c.JSON(http.StatusInternalServerError, response)
 			return
 		}
 	}
 
-	users := paginatedUsers.Users
-	paginationData := paginatedUsers.Pagination
-	response := utils.GetResponse(users, paginationData, http.StatusOK, "Success")
+	response := utils.GetResponse(paginatedUsers.Users, paginatedUsers.Pagination, http.StatusOK, "Success")
 	c.JSON(http.StatusOK, response)
 }
 
@@ -53,7 +51,7 @@ func GetUser(c *gin.Context) {
 
 	user, err := repositories.GetUser(id)
 	if err != nil {
-		response := utils.GetResponse(nil, nil, http.StatusInternalServerError, err.Error())
+		response := utils.GetErrorResponse(http.StatusInternalServerError, err.Error())
 		c.JSON(http.StatusInternalServerError, response)
 		return
 	}
@@ -66,26 +64,26 @@ func CreateUser(c *gin.Context) {
 	username := c.PostForm("username")
 
 	if len(username) < 3 {
-		response := utils.GetResponse(nil, nil, http.StatusBadRequest, "Username must be at least 3 characters long")
+		response := utils.GetErrorResponse(http.StatusBadRequest, "Username must be at least 3 characters long")
 		c.JSON(http.StatusBadRequest, response)
 		return
 	}
 
 	existing_user, err := repositories.GetUserByUsername(username)
 	if err != nil {
-		response := utils.GetResponse(nil, nil, http.StatusInternalServerError, "")
+		response := utils.GetErrorResponse(http.StatusInternalServerError, "")
 		c.JSON(http.StatusInternalServerError, response)
 		return
 	}
 	if existing_user != nil {
-		response := utils.GetResponse(nil, nil, http.StatusBadRequest, "A user with this username already exists")
+		response := utils.GetErrorResponse(http.StatusBadRequest, "A user with this username already exists")
 		c.JSON(http.StatusBadRequest, response)
 		return
 	}
 
 	user, err := repositories.CreateUser(username)
 	if err != nil {
-		response := utils.GetResponse(nil, nil, http.StatusInternalServerError, "")
+		response := utils.GetErrorResponse(http.StatusInternalServerError, "")
 		c.JSON(http.StatusInternalServerError, response)
 		return
 	}
